+++ src/pages/Encounters/tabs/EncounterNotesTab.tsx
@@ -44,7 +44,7 @@ import {
 
 import { Avatar } from "@/components/Common/Avatar";
 import Loading from "@/components/Common/Loading";
-import { MessageSkeleton } from "@/components/Common/UseSkeletons";
+import { MessageSkeleton } from "@/components/Common/SkeletonComponents.tsx";
 
 import useAuthUser from "@/hooks/useAuthUser";
