+++ src/pages/Organization/components/OrganizationLayout.tsx
@@ -15,7 +15,7 @@ import {
 import { Menubar, MenubarMenu, MenubarTrigger } from "@/components/ui/menubar";
 
 import Page from "@/components/Common/Page";
-import { OrganizationIndexSkeleton } from "@/components/Common/UseSkeletons";
+import { OrganizationIndexSkeleton } from "@/components/Common/SkeletonComponents.tsx";
 
 import query from "@/Utils/request/query";
 import { usePermissions } from "@/context/PermissionContext";
